Factor out bootloader handlers to a separete file
